requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
pandas==3.0.1
numpy==2.4.3
requests==2.33.0
supabase==2.28.3
python-dotenv==1.2.2
pydantic==2.12.5
anthropic==0.86.0
pyarrow==23.0.1
httpx==0.28.1
